pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cara-analytics"
version = "0.1.0"
description = "Cara Analytics - Plasma donation data analysis"
authors = [{name = "Cara Team"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "marimo",
    "deepface",
    "insightface", 
    "onnxruntime-gpu",
    "tf-keras",
    "pandas",
    "numpy",
    "matplotlib",
    "seaborn",
    "scikit-learn",
    "sqlalchemy",
    "psycopg2-binary",
    "pymysql",
    "boto3",
    "botocore", 
    "pyarrow",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "ipywidgets",
    "geopandas",
    "hmmlearn",
    "xgboost",
    "openai",
    "tabulate",
    "sqlglot",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock",
    "pytest-cov",
    "black",
    "flake8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["lib*", "locations*"]
exclude = ["experiment*", "data_ingestion*", "face_features_app*", ".git*", ".venv*"]

[tool.setuptools.package-dir]
"" = "."